@@ -15,22 +15,32 @@
             "homepage": "https://github.com/FriendsOfSymfony/oauth2-php/contributors"
         }
     ],
+
     "require": {
-        "php": ">=5.3.2",
-        "symfony/http-foundation": "~2.0"
+        "php": "^5.5.9 || ^7.0.8 || ^7.1.3 || ^7.2.5 || ^8.0.0",
+        "symfony/http-foundation": "~3.0|~4.0|~5.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "phpunit/phpunit": "^5.0 || ^6.0 || ^8.0 || ^9.0.0"
     },
+
     "autoload": {
-        "psr-4": { "OAuth2\\": "lib/" }
+        "psr-4": {
+            "OAuth2\\": "lib/"
+        },
+        "exclude-from-classmap": [
+            "/tests/"
+        ]
     },
     "autoload-dev": {
-        "psr-4": { "OAuth2\\Tests\\": "tests/" }
+        "psr-4": {
+            "OAuth2\\Tests\\": "tests/"
+        }
     },
+
     "extra": {
         "branch-alias": {
-            "dev-master": "1.1.x-dev"
+            "dev-master": "1.2.x-dev"
         }
     }
 }
