@@ -2,6 +2,8 @@
 
 namespace OAuth2;
 
+use Symfony\Component\HttpFoundation\Response;
+
 /**
  * Redirect the end-user's user agent with error message.
  *
@@ -19,6 +21,13 @@ class OAuth2RedirectException extends OAuth2ServerException
     protected $redirectUri;
 
     /**
+     * Parameters are added into 'query' or 'fragment'
+     *
+     * @var string
+     */
+    protected $method;
+
+    /**
      * @param string $redirectUri      An absolute URI to which the authorization server will redirect the user-agent to when the end-user authorization step is completed.
      * @param string $error            A single error code as described in Section 4.1.2.1
      * @param string $errorDescription (optional) A human-readable text providing additional information, used to assist in the understanding and resolution of the error occurred.
@@ -28,10 +37,11 @@ class OAuth2RedirectException extends OAuth2ServerException
      *
      * @ingroup oauth2_error
      */
-    public function __construct($redirectUri, $error, $errorDescription = null, $state = null)
+    public function __construct($redirectUri, $error, $errorDescription = null, $state = null, $method = OAuth2::TRANSPORT_QUERY)
     {
-        parent::__construct(OAuth2::HTTP_FOUND, $error, $errorDescription);
+        parent::__construct(Response::HTTP_FOUND, $error, $errorDescription);
 
+        $this->method = $method;
         $this->redirectUri = $redirectUri;
         if ($state) {
             $this->errorData['state'] = $state;
@@ -47,7 +57,7 @@ class OAuth2RedirectException extends OAuth2ServerException
      */
     public function getResponseHeaders()
     {
-        $params = array('query' => $this->errorData);
+        $params = array($this->method => $this->errorData);
 
         return array(
             'Location' => $this->buildUri($this->redirectUri, $params),
